change cluster selection epsilon to a flot

+++ app/src/pages/embedding/Embedding.tsx
@@ -77,7 +77,7 @@ const EmbeddingUMAPQuery = graphql`
     $nSamples: Int!
     $minClusterSize: Int!
     $clusterMinSamples: Int!
-    $clusterSelectionEpsilon: Int!
+    $clusterSelectionEpsilon: Float!
   ) {
     embedding: node(id: $id) {
       ... on EmbeddingDimension {
